fixed as midas color, disable close button, added vertical margin
